requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.3
pytest==8.0.0
python-multipart==0.0.9
markdown==3.5.2
aiohttp==3.9.3
tenacity==8.2.3
asyncio==3.4.3
jinja2==3.1.3
python-jose==3.3.0
passlib==1.7.4